fix head event
